# public/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Scrum API",
        "description": "Scrum API - For Adding Questions and answers",
        "contact": {
            "email": "roggepatrick@googlemail.com"
        },
        "version": "1.0.1"
    },
    "paths": {
...
        "question/read": {
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "List of questions and answers",
                "operationId": "App\\Http\\Controllers\\QuestionController::read",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "query",
                        "description": "JWT access token.",
                        "required": true,
                        "schema": {
                            "type": "header"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/../../Models/Answers"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}